# backend/docs/swagger.yaml
basePath: /
definitions:
...
  admin.DeprecateModuleVersionRequest:
    properties:
      message:
        type: string
    type: object
...
  admin.MessageResponse:
    properties:
      message:
        type: string
    type: object
  admin.ModuleDetailResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      download_count:
        type: integer
      id:
        type: string
      name:
        type: string
      namespace:
        type: string
      source:
        type: string
      system:
        type: string
      updated_at:
        type: string
      versions:
        items:
          $ref: '#/definitions/admin.ModuleVersionItem'
        type: array
    type: object
...
  admin.ModuleVersionItem:
    properties:
      created_at:
        type: string
      deprecated:
        type: boolean
      deprecated_at: {}
      deprecation_message: {}
      download_count:
        type: integer
      id:
        type: string
      version:
        type: string
    type: object
...
  models.Module:
    properties:
      created_at:
        type: string
      created_by:
        description: User ID who created this module
        type: string
      created_by_name:
        description: Joined fields (not stored in modules table)
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      namespace:
        type: string
      organization_id:
        type: string
      source:
        type: string
      system:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: admin@your-registry.example.com
    name: Registry Support
  description: Complete Terraform Module and Provider Registry with SCM integration,
    mirrors, and storage backends
  license:
    name: Apache-2.0
  termsOfService: https://your-registry.example.com/terms
  title: Terraform Registry API
  version: 1.0.0
paths:
...
  /api/v1/admin/modules/create:
    post:
      consumes:
      - application/json
      description: Create a module record without a version file. Used by the SCM
        publishing flow. Requires modules:publish scope.
      parameters:
      - description: namespace, name, system, description (optional)
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Module already exists (returned as-is)
          schema:
            $ref: '#/definitions/models.Module'
        "201":
          description: Module created
          schema:
            $ref: '#/definitions/models.Module'
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create module record
      tags:
      - Modules
...
  /api/v1/modules/{namespace}/{name}/{system}:
    delete:
      description: Delete a module and all its versions, including files in storage.
        Requires modules:delete scope.
      parameters:
      - description: Module namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Module name
        in: path
        name: name
        required: true
        type: string
      - description: Target system (e.g. aws, azurerm)
        in: path
        name: system
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Module not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete module
      tags:
      - Modules
    get:
      description: Retrieve a module with all its versions, download counts, and metadata.
        No authentication required; authentication is optional and provides user context.
      parameters:
      - description: Module namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Module name
        in: path
        name: name
        required: true
        type: string
      - description: Target system (e.g. aws, azurerm)
        in: path
        name: system
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.ModuleDetailResponse'
        "404":
          description: Module not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get module
      tags:
      - Modules
  /api/v1/modules/{namespace}/{name}/{system}/versions/{version}:
    delete:
      description: Delete a specific version of a module, including its file in storage.
        Requires modules:delete scope.
      parameters:
      - description: Module namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Module name
        in: path
        name: name
        required: true
        type: string
      - description: Target system (e.g. aws, azurerm)
        in: path
        name: system
        required: true
        type: string
      - description: Semantic version (e.g. 1.2.3)
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Module or version not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete module version
      tags:
      - Modules
  /api/v1/modules/{namespace}/{name}/{system}/versions/{version}/deprecate:
    delete:
      description: Remove the deprecated status from a module version. Requires modules:publish
        scope.
      parameters:
      - description: Module namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Module name
        in: path
        name: name
        required: true
        type: string
      - description: Target system (e.g. aws, azurerm)
        in: path
        name: system
        required: true
        type: string
      - description: Semantic version (e.g. 1.2.3)
        in: path
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Module or version not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Undeprecate module version
      tags:
      - Modules
    post:
      consumes:
      - application/json
      description: Mark a specific module version as deprecated with an optional message.
        Requires modules:publish scope.
      parameters:
      - description: Module namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Module name
        in: path
        name: name
        required: true
        type: string
      - description: Target system (e.g. aws, azurerm)
        in: path
        name: system
        required: true
        type: string
      - description: Semantic version (e.g. 1.2.3)
        in: path
        name: version
        required: true
        type: string
      - description: Optional deprecation message
        in: body
        name: body
        schema:
          $ref: '#/definitions/admin.DeprecateModuleVersionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Module or version not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Deprecate module version
      tags:
      - Modules
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: '"JWT token or API key. For JWT: ''Bearer {token}''. For API Key:
      ''Bearer {api_key}''"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"